remove uses of macroquad vector types from Simulation

+++ src/renderer.rs
@@ -15,20 +15,20 @@ pub fn render(camera_target: Vec2, zoom: f32, sim: &Simulation) {
     });
 
     let grid_size = 100.0;
-    let n = 1 + (WORLD_SIZE / grid_size) as i32;
+    let n = 1 + (WORLD_SIZE as f32 / grid_size) as i32;
     for i in -(n / 2)..(n / 2 + 1) {
         shapes::draw_line(
             (i as f32) * grid_size,
-            -WORLD_SIZE / 2.0,
+            (-WORLD_SIZE as f32) / 2.0,
             (i as f32) * grid_size,
-            WORLD_SIZE / 2.0,
+            (WORLD_SIZE as f32) / 2.0,
             1.0,
             color::GREEN,
         );
         shapes::draw_line(
-            -WORLD_SIZE / 2.0,
+            (-WORLD_SIZE as f32) / 2.0,
             (i as f32) * grid_size,
-            WORLD_SIZE / 2.0,
+            (WORLD_SIZE as f32) / 2.0,
             (i as f32) * grid_size,
             1.0,
             color::GREEN,
@@ -36,7 +36,7 @@ pub fn render(camera_target: Vec2, zoom: f32, sim: &Simulation) {
     }
 
     {
-        let v = -WORLD_SIZE / 2.0;
+        let v = -WORLD_SIZE as f32 / 2.0;
         shapes::draw_line(-v, -v, v, -v, 1.0, color::RED);
         shapes::draw_line(-v, v, v, v, 1.0, color::RED);
         shapes::draw_line(-v, -v, -v, v, 1.0, color::RED);
